Add Event::is_eager API
This API allows the event to inform the driver if it is "eager" or not.
The driver can determine whether to eagerly submit this event by calling
io_uring_enter based on this information.
