+++ src/submission.rs
@@ -48,7 +48,8 @@ impl<E: Event, S: Submit> Submission<E, S> {
 
     #[inline(always)]
     unsafe fn try_submit(mut self: Pin<&mut Self>, ctx: &mut Context<'_>) {
-        match self.as_mut().driver().poll_submit(ctx) {
+        let (event, driver) = self.as_mut().event_and_driver();
+        match driver.poll_submit(ctx, event.is_eager()) {
             Poll::Ready(result)  => {
                 let _ = result; // TODO figure out how to handle this result
                 Pin::get_unchecked_mut(self).state = State::Submitted;
@@ -70,11 +71,6 @@ impl<E: Event, S: Submit> Submission<E, S> {
             this.completion.as_mut().set_waker(ctx.waker().clone());
             Poll::Pending
         }
-    }
-
-    #[inline(always)]
-    fn driver(self: Pin<&mut Self>) -> Pin<&mut S> {
-        unsafe { Pin::map_unchecked_mut(self, |this| &mut *this.driver) }
     }
 
     #[inline(always)]
